+++ cmake/yup_standalone.cmake
@@ -26,17 +26,19 @@ function (yup_standalone_app)
         # Globals
         TARGET_NAME TARGET_VERSION TARGET_CONSOLE TARGET_IDE_GROUP TARGET_APP_NAMESPACE TARGET_ICON TARGET_WHEEL TARGET_CXX_STANDARD
         # Emscripten
-        INITIAL_MEMORY MAXIMUM_MEMORY PTHREAD_POOL_SIZE STACK_SIZE CUSTOM_PLIST CUSTOM_SHELL)
+        INITIAL_MEMORY MAXIMUM_MEMORY PTHREAD_POOL_SIZE STACK_SIZE CUSTOM_PLIST CUSTOM_SHELL ENABLE_EMSCRIPTEN_WEBGPU ENABLE_EMSCRIPTEN_GL_DEBUGGING)
     set (multi_value_args
         # Globals
         DEFINITIONS COMPILE_OPTIONS MODULES SOURCES LINK_OPTIONS
         # Emscripten
-        PRELOAD_FILES)
+        PRELOAD_FILES EMSCRIPTEN_LINK_OPTIONS)
 
     cmake_parse_arguments (YUP_ARG "${options}" "${one_value_args}" "${multi_value_args}" ${ARGN})
 
     _yup_set_default (YUP_ARG_TARGET_CXX_STANDARD 20)
     _yup_set_default (YUP_ARG_TARGET_ICON "${CMAKE_CURRENT_FUNCTION_LIST_DIR}/resources/app-icon.png")
+    _yup_set_default (YUP_ARG_ENABLE_EMSCRIPTEN_WEBGPU OFF)
+    _yup_set_default (YUP_ARG_ENABLE_EMSCRIPTEN_GL_DEBUGGING OFF)
 
     set (target_name "${YUP_ARG_TARGET_NAME}")
     set (target_version "${YUP_ARG_TARGET_VERSION}")
@@ -204,29 +206,44 @@ function (yup_standalone_app)
             -pthread
             -Wno-pthreads-mem-growth
             -sWASM=1
+            #-sASYNCIFY=1
             -sWASM_WORKERS=1
             -sAUDIO_WORKLET=1
-            -sPTHREAD_POOL_SIZE=${YUP_ARG_PTHREAD_POOL_SIZE}
-            -sSTACK_SIZE=${YUP_ARG_STACK_SIZE}
             -sSHARED_MEMORY=1
             -sALLOW_MEMORY_GROWTH=1
-            -sINITIAL_MEMORY=${YUP_ARG_INITIAL_MEMORY}
-            $<$<BOOL:${YUP_ARG_MAXIMUM_MEMORY}>:-sMAXIMUM_MEMORY=${YUP_ARG_MAXIMUM_MEMORY}>
             -sASSERTIONS=1
-            # -sGL_ASSERTIONS=1
-            # -sGL_DEBUG=1
+            -sEXIT_RUNTIME=1
             -sDISABLE_EXCEPTION_CATCHING=0
             -sERROR_ON_UNDEFINED_SYMBOLS=1
             -sSTACK_OVERFLOW_CHECK=2
+            -sSTACK_SIZE=${YUP_ARG_STACK_SIZE}
+            -sINITIAL_MEMORY=${YUP_ARG_INITIAL_MEMORY}
+            $<$<BOOL:${YUP_ARG_MAXIMUM_MEMORY}>:-sMAXIMUM_MEMORY=${YUP_ARG_MAXIMUM_MEMORY}>
+            -sPTHREAD_POOL_SIZE=${YUP_ARG_PTHREAD_POOL_SIZE}
             -sFORCE_FILESYSTEM=1
-            -sEXIT_RUNTIME=1
             -sNODERAWFS=0
             -sWASMFS=1
             -sFETCH=1
-            #-sASYNCIFY=1
             -sEXPORTED_RUNTIME_METHODS=ccall,cwrap
             -sDEFAULT_LIBRARY_FUNCS_TO_INCLUDE='$dynCall'
             --shell-file=${YUP_ARG_CUSTOM_SHELL})
+
+        if (YUP_ARG_ENABLE_EMSCRIPTEN_GL_DEBUGGING)
+            list (APPEND additional_link_options
+                -sGL_ASSERTIONS=1
+                -sGL_DEBUG=1)
+        endif()
+
+        if (YUP_ARG_ENABLE_EMSCRIPTEN_WEBGPU)
+            list (APPEND additional_definitions
+                RIVE_WEBGPU=2)
+
+            list (APPEND additional_options
+                --use-port=emdawnwebgpu)
+
+            list (APPEND additional_link_options
+                --use-port=emdawnwebgpu)
+        endif()
 
         foreach (preload_file IN ITEMS ${YUP_ARG_PRELOAD_FILES})
             list (APPEND additional_link_options "--preload-file=${preload_file}")